Animated sprites and Camera

+++ app/src/main/java/fuck/it/attack/joystick/JoystickMovedListener.java
@@ -2,7 +2,7 @@ package fuck.it.attack.joystick;
 
 public interface JoystickMovedListener {
 
-	void onMoved(int pan, int tilt);
+	void onMoved(float pan, float tilt);
 	void onReleased();
 
 }
